little enhancemente
git-svn-id: http://svn.zentyal.org/zentyal/trunk@5312 866bc4c1-54da-0310-a958-9d5a9a0e8d04

+++ client/ebox/src/EBox/Test/Class.pm
@@ -19,6 +19,7 @@ sub _testStubsForFrameworkModules :  Test(startup) {
 
 sub cleanTestDir  : Test(startup) {
     my $testDir = testDir();
+    return if !defined $testDir;
 
     system "rm -rf $testDir"; 
     ($? == 0) or die "Can not clean test dir $testDir";
